try frobenius with floats

+++ src-gen/hlpp19/frobenius-n-1-g-1/CPU-MPMD/include/frobenius-n-1-g-1_0.hpp
@@ -1,3 +1,3 @@
 #pragma once
 
-extern mkt::DMatrix<double> as;
+extern mkt::DMatrix<float> as;
